An attempt to integrate firmata.
The point was to implement a remote control for the sketch using
a generic Firmata implementation. The problem is that the
firmata implementation is way more complicated than necessary for
our purposes, and I have yet to get the app to connect to this
implementation. Saving it, but letting it go.